APIs/azure.com/network/2015-05-01-preview/python-aiohttp-client/tests/test_public_ip_addresses_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.public_ip_address import PublicIpAddress
from openapi_server.models.public_ip_address_list_result import PublicIpAddressListResult


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("Connexion does not support multiple consumes. See https://github.com/zalando/connexion/pull/760")
async def test_public_ip_addresses_create_or_update(client):
    """Test case for public_ip_addresses_create_or_update

    
    """
    parameters = {"etag":"etag","properties":{"dnsSettings":{"reverseFqdn":"reverseFqdn","fqdn":"fqdn","domainNameLabel":"domainNameLabel"},"ipConfiguration":{"id":"id"},"resourceGuid":"resourceGuid","idleTimeoutInMinutes":0,"ipAddress":"ipAddress","provisioningState":"provisioningState","publicIPAllocationMethod":"Static"}}
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='PUT',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/publicIPAddresses/{public_ip_address_name}'.format(resource_group_name='resource_group_name_example', public_ip_address_name='public_ip_address_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        json=parameters,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_public_ip_addresses_delete(client):
    """Test case for public_ip_addresses_delete

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/publicIPAddresses/{public_ip_address_name}'.format(resource_group_name='resource_group_name_example', public_ip_address_name='public_ip_address_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_public_ip_addresses_get(client):
    """Test case for public_ip_addresses_get

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/publicIPAddresses/{public_ip_address_name}'.format(resource_group_name='resource_group_name_example', public_ip_address_name='public_ip_address_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_public_ip_addresses_list(client):
    """Test case for public_ip_addresses_list

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/resourceGroups/{resource_group_name}/providers/Microsoft.Network/publicIPAddresses'.format(resource_group_name='resource_group_name_example', subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_public_ip_addresses_list_all(client):
    """Test case for public_ip_addresses_list_all

    
    """
    params = [('api-version', 'api_version_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/subscriptions/{subscription_id}/providers/Microsoft.Network/publicIPAddresses'.format(subscription_id='subscription_id_example'),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

